pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zeer"
version = "0.2.0"
description = "Agentic AI CLI with tool calling and extensible skills system"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "M Aliyan", email = "mahunxai@gmail.com"}
]
keywords = ["ai", "agent", "cli", "tool-calling", "agentic", "openai", "gemini", "claude", "agentskills"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "requests>=2.31.0",
    "inquirer>=3.1.3",
    "colorama>=0.4.6",
    "prompt_toolkit>=3.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/muhdaliyan/zeer"
Documentation = "https://github.com/muhdaliyan/zeer#readme"
Repository = "https://github.com/muhdaliyan/zeer"
Issues = "https://github.com/muhdaliyan/zeer"

[project.scripts]
zeer = "src.main:main"

[tool.setuptools]
packages = ["src", "src.providers", "skills", "skills.code-helper", "skills.file-operations",
 "skills.frontend-design", "skills.pdf", "skills.pptx", "skills.text-processing", "skills.webapp-testing",
 "skills.analyzing-time-series", "skills.analyzing-time-series"]
include-package-data = true

[tool.setuptools.package-data]
skills = ["**/*.md", "**/*.py"]